pkg/emulatedclients/key/generator/generator-always-refresh_test.go
package generator

import (
	"fmt"
	"github.com/anacrolix/torrent/metainfo"
	"github.com/anacrolix/torrent/tracker"
	"github.com/stretchr/testify/assert"
	"gopkg.in/yaml.v2"
	"testing"
)

func TestUnmarshalAlwaysRefreshGenerator(t *testing.T) {
	yamlString := `---
type: ALWAYS_REFRESH
algorithm:
  type: NUM_RANGE_ENCODED_AS_HEXADECIMAL
  min: 1
  max: 2
`
	generator := &KeyGenerator{}
	err := yaml.Unmarshal([]byte(yamlString), generator)
	if err != nil {
		t.Fatalf("Failed to unmarshall: %+v", err)
	}
	_ = generator.afterPropertiesSet()
	assert.IsType(t, &AlwaysRefreshGenerator{}, generator.IKeyGenerator)
}

func TestGenerateAlwaysRefresh(t *testing.T) {
	generator := &AlwaysRefreshGenerator{}
	_ = generator.afterPropertiesSet()

	dumbAlg := &DumbAlgorithm{}
	for i := 0; i < 500; i++ {
		infoHash := metainfo.NewHashFromHex(fmt.Sprintf("%dAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA", i)[0:40])
		_ = generator.get(dumbAlg, infoHash, tracker.None)
	}

	assert.Equal(t, 500, dumbAlg.counter, "Should have been called 500 times")
}
